WIP: Semi-working switch between environments
Still needs some tweaking in Cura and our plugin here.
However, good enough to select faces, find the closest node. Process the mesh and look for the contour with the node included.
Which will be our matched surface by the way.

Need to go back with this WIP back to the Ultimaker team, as believe that I should be able to see faces already.
But I get always None back :(

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com> (github: thopiekar)