v19.0.5.0: Requirements: More UI tweaks and adding Qt signals
Now the UI distrubutes the default values when Cura is fully set up.
That is done using Qt Signals, to which our Extension plugin can connect to.
This way the Extension plugin can update the own cached values, which are later bundled to a 3MF and sent to the cloud.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com> (github: thopiekar)