tried to clean up the force definition
the old force definition was hard to keep track of because all of the info (magnitude, normal vector, push/pull) was stored into a single vector. So managing the state was difficult when you only wanted to change one attribute, such as the push/pull. I re-worked it into a small class and tried not to disturb too much else, because this will likely get re-hauled when multiple anchors/loads is implemented. This also fixed some problems with updating the state of the force vector that wouldn't propagate to the Smart Slice job